requirements.txt
langchain
langchain-community
langchain-text-splitters
chromadb
sentence-transformers
gitpython
streamlit
python-dotenv
google-generativeai